feat: tests and refactoring into mainnet folder
