+++ contracts/mainnet/mocks/CryptoRegistry.vy
