+++ contracts/mainnet/mocks/StableFactory.vy
