+++ contracts/mainnet/mocks/StableRegistry.vy
