+++ scripts/mainnet/print_missing.py
