+++ tests/mainnet/abis/CurvePool.json
