+++ tests/mainnet/abis/StableRegistry.json
